+++ flowman-core/src/main/scala/com/dimajix/flowman/history/JdbcStateStore.scala
@@ -29,8 +29,10 @@ import slick.jdbc.H2Profile
 import slick.jdbc.MySQLProfile
 import slick.jdbc.PostgresProfile
 
+import com.dimajix.flowman.execution.JobToken
 import com.dimajix.flowman.execution.Phase
 import com.dimajix.flowman.execution.Status
+import com.dimajix.flowman.execution.TargetToken
 import com.dimajix.flowman.model.JobInstance
 import com.dimajix.flowman.model.TargetInstance