+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/target/LocalTarget.scala
@@ -46,7 +46,7 @@ import com.dimajix.flowman.model.TargetInstance
 case class LocalTarget(
     instanceProperties:Target.Properties,
     mapping:MappingOutputIdentifier,
-    filename:String,
+    path:String,
     encoding:String,
     header:Boolean,
     newline:String,
@@ -66,7 +66,7 @@ case class LocalTarget(
             namespace.map(_.name).getOrElse(""),
             project.map(_.name).getOrElse(""),
             name,
-            Map("filename" -> filename)
+            Map("path" -> path)
         )
     }
 
@@ -81,7 +81,7 @@ case class LocalTarget(
       * @return
       */
     override def provides(phase: Phase) : Set[ResourceIdentifier] = Set(
-        ResourceIdentifier.ofLocal(new Path(filename))
+        ResourceIdentifier.ofLocal(new Path(path))
     )
 
     /**
@@ -101,14 +101,14 @@ case class LocalTarget(
       * @param executor
       */
     override def build(executor:Executor) : Unit = {
-        logger.info(s"Writing mapping '${this.mapping}' to local file '$filename'")
+        logger.info(s"Writing mapping '${this.mapping}' to local file '$path'")
 
         val mapping = context.getMapping(this.mapping.mapping)
         val dfIn = executor.instantiate(mapping, this.mapping.output)
         val cols = if (columns.nonEmpty) columns else dfIn.columns.toSeq
         val dfOut = dfIn.select(cols.map(c => dfIn(c).cast(StringType)):_*)
 
-        val outputFile = new File(filename)
+        val outputFile = new File(path)
         outputFile.getParentFile.mkdirs()
         outputFile.createNewFile
         val outputStream = new FileOutputStream(outputFile)
@@ -143,9 +143,9 @@ case class LocalTarget(
     override def verify(executor: Executor) : Unit = {
         require(executor != null)
 
-        val file = executor.fs.local(filename)
+        val file = executor.fs.local(path)
         if (!file.exists()) {
-            logger.error(s"Verification of target '$identifier' failed - local file '$filename' does not exist")
+            logger.error(s"Verification of target '$identifier' failed - local file '$path' does not exist")
             throw new VerificationFailedException(identifier)
         }
     }
@@ -158,9 +158,9 @@ case class LocalTarget(
     override def truncate(executor: Executor): Unit = {
         require(executor != null)
 
-        val outputFile = new File(filename)
+        val outputFile = new File(path)
         if (outputFile.exists()) {
-            logger.info(s"Cleaning local file '$filename'")
+            logger.info(s"Cleaning local file '$path'")
             outputFile.delete()
         }
     }